requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
azure-cognitiveservices-speech==1.40.0
certifi==2025.4.26
click==8.2.1
distro==1.9.0
fastapi[standard]~=0.115.12
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.82.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2

PyJWT~=2.10.1
pytest==7.4.0
pytest-asyncio==0.21.1
bcrypt~=4.3.0
passlib~=1.7.4
cryptography~=45.0.4
requests~=2.32.4

botocore~=1.35.0
boto3~=1.35.0
python-multipart~=0.0.9
pydub==0.25.1

numpy>=1.26.0,<2.0.0
pandas>=2.2.0,<2.3.0